BUG:
There is a very strange bug I am having trouble isolating. When developing in IDEA the bug does not exist, when I run Maven 'package' to create a standalone executable jar, the bug exists.

The bug is that when pressing the pause button, the program will pause, but pressing again with not unpause. It works in dev but not packaged in the jar.

Also, the bug seems to only occur when wait time is longer. I could not produce the bug with a 10second wait time, but can with a 40 second wait time.

When I print out the status of 'paused' in the main loop, the bug is no longer able to be reproduced in the packaged jar.

I would say the issue is closed now but I didn't really figure out why that bug was happening...